% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.R
\name{arm_id}
\alias{arm_id}
\title{Unique name for arm}
\usage{
arm_id(task, splitting = FALSE)
}
\value{
A character vector of length 1, the arm_name
}
\description{
Create a character name for a particular arm
}
\details{
NOTE THIS SEEMS TO DUPLICATE taskarm_name??
}
