% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.R
\name{arms_factorial}
\alias{arms_factorial}
\title{Build factorial arms}
\usage{
arms_factorial(...)
}
\arguments{
\item{...}{A number of parameters}
}
\description{
Adds arms of ALL combinations of the supplied variables as arms
}
\examples{
# will run 4 arms, corresponding to each combination
arms_factorial(seed = c(21, 42), type = c("fast", "slow"))
}
