% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat_df}
\alias{cat_df}
\title{Concatenate and print a data.frame}
\usage{
cat_df(x)
}
\arguments{
\item{x}{A data frame}
}
\description{
Output a data.frame object using cat. This is useful so that raw bash
calls (for colours etc) are printed correctly
}
