% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_dir_retain_dates}
\alias{copy_dir_retain_dates}
\title{Copy directories, retaining last modified dates of files}
\usage{
copy_dir_retain_dates(from, to)
}
\arguments{
\item{from}{Directory to copy}

\item{to}{Directory to copy to}
}
\description{
file.copy does this in principle, but it does not seem to work on Linux
}
