% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_git_toplevel}
\alias{dir_git_toplevel}
\title{Get the toplevel git directory}
\usage{
dir_git_toplevel(dir = getwd(), check = TRUE)
}
\arguments{
\item{dir}{A directory, by default the current working directory}

\item{check}{Logical, if TRUE no test for whether dir is inside a git work
tree}
}
\value{
The path to the toplevel git directory
}
\description{
Get the toplevel git directory
}
