% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_run_branch_toplevel}
\alias{dir_run_branch_toplevel}
\title{Toplevel directory of the (branch-specific) run directory}
\usage{
dir_run_branch_toplevel(dir = getwd(), check = TRUE)
}
\arguments{
\item{dir}{A directory}

\item{check}{Logical, if FALSE no test for whether dir is inside a git work
tree}
}
\value{
Path to the git toplevel for the run directory corresponding to dir
}
\description{
If passed ~/analyses/project/folder/, where ~/analyses/project is the git
toplevel folder, will return
}
\details{
mngr_option_dir_run("$GIT_TOPLEVEL")/$GIT_BRANCH/

If mngr_option_dir_run($GIT_TOPLEVEL)=$GIT_TOPLEVEL, then just returns what
was supplied
}
