% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_path}
\alias{file_path}
\title{Build a file path}
\usage{
file_path(file, arm = arm_id(.task, splitting = splitting),
  directory = "etc", extension = "", splitting = FALSE)
}
\arguments{
\item{file}{file name}

\item{arm}{make arm-specific}

\item{directory}{directory name}

\item{extension}{file extension}

\item{splitting}{A logical, should the split arm names be combined in a
single part}
}
\value{
A character vector of length 1, the file path
}
\description{
Build a file path
}
