% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mngr.R
\name{find_mngrfile}
\alias{find_mngrfile}
\title{Find Mngrfile}
\usage{
find_mngrfile(dir)
}
\arguments{
\item{dir}{a path to a directory}
}
\description{
Find the Mngrfile in the supplied directory or the parent directory
}
