% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{get_jobids}
\alias{get_jobids}
\title{Extract Job ID from squeue output}
\usage{
get_jobids(pattern = NULL)
}
\arguments{
\item{pattern}{A regular expression of Job Names to extract the Job ID of}
}
\value{
A vector of Job IDs
}
\description{
Extract Job ID from squeue output
}
