% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_abbrev_ref}
\alias{git_abbrev_ref}
\title{Abbreviated name for HEAD commit in git repository}
\usage{
git_abbrev_ref(dir = getwd(), base_only = TRUE)
}
\arguments{
\item{dir}{A directory}

\item{base_only}{Logical, should only the first part of a foo/bar branch
name be returned}
}
\value{
A character with the abbrev-ref
}
\description{
Uses git rev-parse --abbrev-ref to get the branch name when appropriate
}
