% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_clone_or_pull}
\alias{git_clone_or_pull}
\title{Clone or pull to the run dir}
\usage{
git_clone_or_pull(dir = getwd())
}
\arguments{
\item{dir}{directory to clone or pull to the corresponding rundir}
}
\description{
Clone or pull to the directory to the corresponding run directory
}
\details{
1. Ensure dir is a git worktree
2. Get the git HEAD name
3. Create the run directory if it does not exist
4. Clone/pull as required
}
