% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_short_sha}
\alias{git_short_sha}
\title{Short SHA for HEAD commit in git repository}
\usage{
git_short_sha(dir = getwd(), length = 4)
}
\arguments{
\item{dir}{A directory}

\item{length}{Number of characters}
}
\value{
A character vector of length 1 containing the short SHA for the HEAD
commit
}
\description{
Uses git rev-parse
}
