% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{\%cores\%}
\alias{\%cores\%}
\title{Set the number of cores for a task}
\usage{
taskname \%cores\% cores
}
\arguments{
\item{taskname}{A task}

\item{cores}{An integer number of cores needed by the task}
}
\description{
Set the number of cores for a task
}
