% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{\%hours\%}
\alias{\%hours\%}
\title{Set the number of hours for a task (walltime)}
\usage{
taskname \%hours\% hours
}
\arguments{
\item{taskname}{A task}

\item{hours}{A number of hours (walltime) that the task will need}
}
\description{
Set the number of hours for a task (walltime)
}
