% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{\%memory\%}
\alias{\%memory\%}
\title{Set the memory requirements for a task}
\usage{
taskname \%memory\% memory
}
\arguments{
\item{taskname}{A task}

\item{memory}{An integer number of megabytes that the task will need}
}
\value{
Invisibly returns the taskname
}
\description{
Set the memory requirements for a task
}
