% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{\%task\%}
\alias{\%task\%}
\title{Create a new task}
\usage{
name \%task\% action
}
\arguments{
\item{name}{task name}

\item{action}{a set of expressions}
}
\description{
Create a new task
}
\author{
RJB Goudie
}
