% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{\%throttle\%}
\alias{\%throttle\%}
\title{Throttle the number of tasks to run simultaneously}
\usage{
taskname \%throttle\% throttle
}
\arguments{
\item{taskname}{A task}

\item{throttle}{The maximum number of task to run simultaneously}
}
\description{
Throttle the number of tasks to run simultaneously
}
