% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{hms}
\alias{hms}
\title{Convert seconds to formatted h:mm:ss string}
\usage{
hms(seconds)
}
\arguments{
\item{seconds}{A numeric number of seconds}
}
\value{
A character vector of the time formatted as "h:mm:ss"
}
\description{
Convert seconds to formatted h:mm:ss string
}
