% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{is_inside_git_work_tree}
\alias{is_inside_git_work_tree}
\title{Check if dir is inside git work tree}
\usage{
is_inside_git_work_tree(dir = getwd())
}
\arguments{
\item{dir}{A directory, by default the current working directory}
}
\value{
A logical indicating whether dir is inside a git work tree. Returns
  FALSE if there are errors or warnings when running rev-parse.
}
\description{
Uses git rev-parse to test whether the supplied directory is a git work
tree.
}
