% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{kill_pattern}
\alias{kill_pattern}
\title{Kill Slurm jobs matching pattern}
\usage{
kill_pattern(shortsha = NULL, user = NULL)
}
\arguments{
\item{shortsha}{The short SHA of the jobs to be killed}

\item{user}{The Slurm user whose jobs they are}
}
\description{
Kill Slurm jobs matching pattern
}
