% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{latest_logs}
\alias{latest_logs}
\title{Summaries the latest R output}
\usage{
latest_logs(filter = FALSE)
}
\arguments{
\item{filter}{If TRUE, then only show failed jobs}
}
\description{
Displays a pretty table of the latest R output
}
