% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{link_to_arm_indicies}
\alias{link_to_arm_indicies}
\title{Merge rout status table with arm data}
\usage{
link_to_arm_indicies(x)
}
\arguments{
\item{x}{A data.frame with "arm___xxx" columns}
}
\description{
Merges rout status table with arm data, so that even when mngr-latest-logs
is run we still know which arm corresponds to which values etc
}
