% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lscheduler.R
\name{lscheduler_add}
\alias{lscheduler_add}
\title{Add to lscheduler}
\usage{
lscheduler_add(cmd, args, dependency, jobid)
}
\arguments{
\item{name}{job name}

\item{action}{a set of expressions}
}
\description{
Add to lscheduler
}
