% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{merge_tables}
\alias{merge_tables}
\title{Merge squeue and rout status tables}
\usage{
merge_tables(squeue_status, rout_df)
}
\arguments{
\item{squeue_status}{A squeue table}

\item{rout_df}{A data.frame of rout}
}
\value{
A merged table
}
\description{
Merge squeue and rout status tables
}
