% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mngr_config}
\alias{mngr_config}
\title{Set configuration options for mngr}
\usage{
mngr_config(...)
}
\arguments{
\item{...}{a list of values to replace default config}
}
\description{
The current options are:
- queue: the Slurm queue to submit to
- output: where the main output should go
- r_logs: where the R logs should be stored
- slurm_logs: where the Slurm logs should be stored
}
