% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mngr_option_dir_output}
\alias{mngr_option_dir_output}
\title{Converting run paths to output paths}
\usage{
mngr_option_dir_output()
}
\value{
The current value of the option
}
\description{
Set the "output" location: this is where all of the plots, rds files, and
logs are stored. You may wish for this to on a separate drive to the R
source, since the output may be large and can be reproduced, so you might
not need to back the results up so carefully. e.g. a 'scratch' disk may
suffice.
}
\details{
This option should be a set to function that converts a path (within the
run directory) to the corresponding path in the results location.

By default, ```mngr_dir_output`` is set so that the results will be saved
within the run directory.
}
