% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mngr_option_dir_r_logs}
\alias{mngr_option_dir_r_logs}
\title{Path to keep all R logs}
\usage{
mngr_option_dir_r_logs()
}
\value{
The current value of the option
}
\description{
This should be a function that converts a runpath to a directory wherer you
wish to store all R logs.
}
\details{
By default, R logs will be stored within the ```mngr_dir_output``` in a
folder called ```r``` within a folder called ```logs```.
}
