% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mngr_option_dir_r_logs_latest}
\alias{mngr_option_dir_r_logs_latest}
\title{Path to keep the latest R logs}
\usage{
mngr_option_dir_r_logs_latest()
}
\value{
The current value of the option
}
\description{
This should be a function that converts a runpath to a directory wherer you
wish to store the latest R logs.
}
\details{
By default, R logs will be stored within the ```mngr_dir_output``` in a
folder called ```r-latest``` within a folder called ```logs```.
}
