% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mngr_option_dir_run}
\alias{mngr_option_dir_run}
\title{Converting paths to runpaths}
\usage{
mngr_option_dir_run()
}
\value{
The function used for converting filepath to the runpath
}
\description{
This option determines the function used to set the path where an analysis
will be run. mngr checks out the git repository to a separate location, when
running the analysis, so that the original working copy of the R files can
be edited while the original analysis is running, without changing what runs.
}
\details{
This option should be set to a function that converts \code{$GIT_TOPLEVEL}
to the equivalant run directory.

By default \code{$HOME/path/to/$GIT_TOPLEVEL/} will be converted to
\code{$HOME/run/$GIT_TOPLEVEL}.
}
