% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mngr_option_dir_state}
\alias{mngr_option_dir_state}
\title{Path to keep the state files}
\usage{
mngr_option_dir_state()
}
\value{
The current value of the option
}
\description{
This should be a function that converts a runpath to a directory wherer you
wish to store the state files.
}
\details{
By default, state files will be stored within the ```mngr_dir_output``` in a
folder called ```state```.
}
