% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mngr_option_global_throttle}
\alias{mngr_option_global_throttle}
\title{Global throttle setting}
\usage{
mngr_option_global_throttle()
}
\value{
The current value of the option
}
\description{
For local scheduler, the maximum number of tasks to run concurrently. The
default is given by \code{\link[parallel]{detectCores}}.
}
\details{
For slurm scheduler, the maximum number of arms of a single task to run
concurrently. The default is 50.
}
