% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mngr_option_scheduler}
\alias{mngr_option_scheduler}
\title{Scheduler used for running R jobs}
\usage{
mngr_option_scheduler()
}
\value{
The current value of the option
}
\description{
This option should be set to either \code{"slurm"} or \code{"local"}.
}
\details{
If \code{"slurm"}, then the Slurm Workload Manager will be used to submit
the jobs to a cluster

If \code{"local"}, then the jobs will be run on the local machine, with
execution controlled by a simple R-based scheduler.

The default is \code{"local"}.
}
