% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mngr_option_slurm_submit_path}
\alias{mngr_option_slurm_submit_path}
\title{Path to the directory containing Slurm submit script}
\usage{
mngr_option_slurm_submit_path()
}
\value{
The current value of the option
}
\description{
This option controls which directory will be searched for Slurm submit
scripts. The default is \code{~/.mngr/slurm/}
}
