% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mngr_option_use_tempfile}
\alias{mngr_option_use_tempfile}
\title{Save RDS data to a tempfile initially}
\usage{
mngr_option_use_tempfile()
}
\value{
The current value of the option
}
\description{
This option determines whether files saved using saveRDS are first saved to
to a tempfile, generated using \code{\link{tempfile}}, and then copied
to the final destination. This makes saving much faster on some very slow
filesystems.
}
