% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor}
\alias{monitor}
\title{Monitor Slurm jobs}
\usage{
monitor(jobs = NULL, logs = NULL, width = NULL)
}
\description{
This function is designed to work from the command line. It accepts two
arguments. The first is a comma-separated list of Slurm Job IDs. The second
is a path to R log directory.
}
