% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor_logs}
\alias{monitor_logs}
\title{Print status of specified jobs}
\usage{
monitor_logs(jobs, dir)
}
\arguments{
\item{jobs}{A character vector of job IDs}

\item{dir}{Path to the directory where the R logs are stored. A character
vector of length 1}
}
\description{
Print status of specified jobs
}
