% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{parse_name}
\alias{parse_name}
\title{Parse job names}
\usage{
parse_name(x, type = "rout")
}
\arguments{
\item{x}{A job name}

\item{type}{Either "rout", "rlatest" or "squeue"}
}
\description{
Parse job names
}
