% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlogs.R
\name{parse_rout_files}
\alias{parse_rout_files}
\title{Parse .Rout files to a data frame}
\usage{
parse_rout_files(paths, type = "rout", filter = FALSE)
}
\arguments{
\item{paths}{Character vector of paths to Rout files}

\item{type}{Passed to parse_name}

\item{filter}{If TRUE, then only show failed jobs

A data.frame with the following columns:

jid, jobname, warnings, errors, internal_run_seconds, final

plus

arm___xyz columns for each arm value}
}
\description{
Parse .Rout files to a data frame
}
