% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{pdf_file}
\alias{pdf_file}
\title{path to pdf file}
\usage{
pdf_file(...)
}
\arguments{
\item{...}{passsed to file_path}
}
\value{
A character vector of length 1, the file path
}
\description{
path to pdf file
}
