% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{png_file}
\alias{png_file}
\title{path to png file}
\usage{
png_file(...)
}
\arguments{
\item{...}{passsed to file_path}
}
\value{
A character vector of length 1, the file path
}
\description{
path to png file
}
