% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.R
\name{prereq_arm_ids}
\alias{prereq_arm_ids}
\title{Arm IDs for the prerequisite tasks}
\usage{
prereq_arm_ids(task)
}
\arguments{
\item{task}{A task name}
}
\value{
A list of arm ids of the current arm of the supplied task
}
\description{
Arm IDs for the prerequisite tasks
}
