% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{pretty_print_arm_values}
\alias{pretty_print_arm_values}
\title{Print the arm values prettily}
\usage{
pretty_print_arm_values(x)
}
\arguments{
\item{x}{A set of arm values}
}
\description{
Print the arm values prettily
}
