% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{pretty_print_time}
\alias{pretty_print_time}
\title{Pretty print the time}
\usage{
pretty_print_time(t)
}
\arguments{
\item{t}{A vector of times}
}
\value{
A character vector of pretty print times
}
\description{
Print times as "XXh XXm XXs"
}
