% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{pretty_success_status}
\alias{pretty_success_status}
\title{Pretty print ticks and crosses}
\usage{
pretty_success_status(finished, success)
}
\arguments{
\item{finished}{A logical vector of finished indicators}

\item{success}{A logiscal vector of success indicators}
}
\value{
A character vector of ticks, crosses and NAs
}
\description{
Returns a vector with ticks for components that have finished successfully,
and crosses for components that finished unsuccessfully
}
