% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{qdel}
\alias{qdel}
\title{A wrapper around Sluerm qdel}
\usage{
qdel(jobids)
}
\arguments{
\item{jobids}{A vector of Slurm Job IDs}
}
\value{
Any errors from qdel will be returned
}
\description{
A wrapper around Sluerm qdel
}
