% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{rds_file}
\alias{rds_file}
\title{path to rds file}
\usage{
rds_file(...)
}
\arguments{
\item{...}{passsed to file_path}
}
\value{
A character vector of length 1, the file path
}
\description{
path to rds file
}
