% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.R
\name{read_rds_merge}
\alias{read_rds_merge}
\title{Load and merge all RDS files}
\usage{
read_rds_merge(...)
}
\arguments{
\item{...}{Passed to rds_file}
}
\value{
A list, each component of which corresponds to a unmerged arm
}
\description{
When a task %merge% a particular arm, then you may wish to load all the
output from the unmerged task.
}
\details{
All files are loaded, and joined into a list
}
