% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfiles.R
\name{rfile_create}
\alias{rfile_create}
\title{Create rfile task}
\usage{
rfile_create(name, action)
}
\arguments{
\item{name}{task name}

\item{action}{a set of expressions (currently ignored?)}
}
\description{
Create rfile task
}
