% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfiles.R
\name{rfiles}
\alias{rfiles}
\title{Add all R files in the current directory as tasks}
\usage{
rfiles()
}
\description{
All files ending .R are added as tasks. Each task takes the name of the file,
minus the file extension
}
