% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mngr.R
\name{run}
\alias{run}
\title{Execute a task}
\usage{
run(name = "default", dry_run = FALSE, debug = FALSE, scheduler = NULL)
}
\arguments{
\item{name}{a task name}

\item{dry_run}{Don't actually execute tasks}

\item{debug}{A logical, indicating whether to show debug messages}

\item{scheduler}{Override the default scheduler}
}
\description{
Actually run a task, and any prerequisites that have changed since the last
run.
}
\details{
The process is as follows:
1. Ensures the current directory is a git work tree
2. Ensures the current directory is a clean git work directory
3. Loads the Mngrfile
4. Invokes the supplied task
}
