% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlogs.R
\name{run_time_seconds}
\alias{run_time_seconds}
\title{Parse the run time of a file}
\usage{
run_time_seconds(path)
}
\arguments{
\item{path}{A path to an Rout log file}
}
\description{
Assumes the runtime appears in green on the final line of the R output
}
