% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{saveRDS}
\alias{saveRDS}
\title{Save an R object to a rds file}
\usage{
saveRDS(..., file)
}
\arguments{
\item{...}{passed to default \code{\link{saveRDS}}}

\item{file}{A filename file name}
}
\description{
Wraps the default \code{\link{saveRDS}} function to allow for handling
softlinking, and using tempfile saving.
}
\details{
If the option \code{mngr_use_tempfile} is TRUE, then files are first saved
to a tempfile, generated using \code{\link{tempfile}}, and then copied
to the final destination. This makes saving much faster on some very slow
filesystems.
}
