% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{sinfo}
\alias{sinfo}
\title{Information about slurm cluster}
\usage{
sinfo(partition = "mrc-bsu-sand,mrc-bsu-tesla")
}
\arguments{
\item{partition}{A character string containing comma-separated Slurm
partitions}
}
\description{
A wrapper around Slurm sinfo
}
