% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{slurm_summary}
\alias{slurm_summary}
\title{Return a brief summary of the Slurm queue status}
\usage{
slurm_summary(squeue_status)
}
\arguments{
\item{squeue_status}{TODO}
}
\description{
Return a brief summary of the Slurm queue status
}
