% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{squeue}
\alias{squeue}
\title{Wrapper around Slurm's squeue function}
\usage{
squeue(jobs = NULL, user = NULL)
}
\arguments{
\item{jobs}{Vector of Slurm Job IDs. NULL returns all jobs}

\item{user}{Vector of usernames. NULL returns all jobs}
}
\value{
A data.frame with the following columns:

jid, partition, jobname, shortsha, user, slurm_status, t_used, nodes,
nodelist_reason, t_left, priority

plus

arm___xyz columns for each arm value
}
\description{
Wrapper around Slurm's squeue function
}
