% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{startup}
\alias{startup}
\title{Startup mngr}
\usage{
startup()
}
\description{
This function set up the .task, .job and .arm variables for non-interactive
jobs, and loads the Mngrfile, so that the arm values are available
}
