% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{system_in_dir}
\alias{system_in_dir}
\title{Evalutate system command in directory}
\usage{
system_in_dir(command, dir, ...)
}
\arguments{
\item{command}{a system call}

\item{dir}{path to directory}

\item{...}{passed to system()}
}
\description{
Wrapper around system to execute a system command in a different
directory to the current one
}
