% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{tiff_file}
\alias{tiff_file}
\title{path to tiff file}
\usage{
tiff_file(...)
}
\arguments{
\item{...}{passsed to file_path}
}
\value{
A character vector of length 1, the file path
}
\description{
path to tiff file
}
