% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_dir}
\alias{with_dir}
\title{Evaluate expression in a particular working directory}
\usage{
with_dir(dir, expr)
}
\arguments{
\item{dir}{path to directory in which to evaluate expression}

\item{expr}{an expression}
}
\description{
Evaluate expression in a particular working directory
}
